pyproject.toml
[project]
name = "hekmatica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4.1",
    "baml-py>=0.86.1,<1.0.0",
    "requests>=2.32.3,<3.0.0",
    "langchain-community>=0.3.20,<1.0.0",
    "langgraph-cli[inmem]>=0.1.81,<1.0.0",
    "langchain-google-genai>=2.1.2,<3.0.0",
    "gradio>=5.25.2,<6.0.0",
    "sentence-transformers>=4.1.0,<5.0.0",
    "faiss-cpu>=1.10.0,<2.0.0",
    "rank-bm25>=0.2.2,<1.0.0",
    "tiktoken>=0.9.0,<1.0.0",
    "loguru>=0.7.3",
    "tavily-python>=0.7.1",
]
